+++ pl_bolts/datamodules/ssl_imagenet_dataloaders.py
@@ -2,7 +2,7 @@ import os
 from torchvision import transforms as transform_lib
 from torch.utils.data import DataLoader
 from pl_bolts.datamodules.bolts_dataloaders_base import BoltDataLoaders
-from pl_bolts.models.self_supervised.amdim.ssl_datasets import UnlabeledImagenet
+from pl_bolts.datamodules.imagenet_dataset import UnlabeledImagenet
 from pl_bolts.transforms.dataset_normalizations import imagenet_normalization
