+++ pl_bolts/models/self_supervised/amdim/datasets.py
@@ -1,8 +1,9 @@
 from torch.utils.data import random_split
 from torchvision.datasets import STL10
 
+from pl_bolts.datamodules.imagenet_dataset import UnlabeledImagenet
 from pl_bolts.models.self_supervised.amdim import transforms as amdim_transforms
-from pl_bolts.models.self_supervised.amdim.ssl_datasets import CIFAR10Mixed, UnlabeledImagenet
+from pl_bolts.models.self_supervised.amdim.ssl_datasets import CIFAR10Mixed
 
 
 class AMDIMPretraining():
